test: move JSON replacement stuff into `testcmd`
