+++ cmd/osv-scanner/internal/testcmd/case.go
@@ -1,14 +1,12 @@
 package testcmd
 
-import "github.com/google/osv-scanner/v2/internal/testutility"
-
 type Case struct {
 	Name string
 	Args []string
 	Exit int
 
 	// ReplaceRules are only used for JSON output
-	ReplaceRules []testutility.JSONReplaceRule
+	ReplaceRules []JSONReplaceRule
 }
 
 func (c Case) isOutputtingJSON() bool {
