+++ cmd/osv-scanner/internal/testcmd/jsonreplace.go
@@ -1,4 +1,4 @@
-package testutility
+package testcmd
 
 import (
 	"strconv"
