+++ cmd/osv-scanner/internal/testcmd/jsonreplace_test.go
@@ -1,4 +1,4 @@
-package testutility
+package testcmd
 
 import (
 	"bytes"
