+++ cmd/osv-scanner/internal/testcmd/run.go
@@ -1,6 +1,8 @@
 package testcmd
 
 import (
+	"bytes"
+	"encoding/json"
 	"testing"
 
 	"github.com/google/osv-scanner/v2/cmd/osv-scanner/internal/cmd"
@@ -28,11 +30,29 @@ func Run(t *testing.T, tc Case) {
 	stdout, stderr := run(t, tc)
 
 	if tc.isOutputtingJSON() {
-		stdout = testutility.NormalizeJSON(t, stdout, tc.ReplaceRules...)
+		stdout = normalizeJSON(t, stdout, tc.ReplaceRules...)
 	}
 
 	testutility.NewSnapshot().MatchText(t, stdout)
 	testutility.NewSnapshot().WithWindowsReplacements(map[string]string{
 		"CreateFile": "stat",
 	}).MatchText(t, stderr)
+}
+
+// normalizeJSON runs the given JSONReplaceRules on the given JSON input and returns the normalized JSON string
+func normalizeJSON(t *testing.T, jsonInput string, jsonReplaceRules ...JSONReplaceRule) string {
+	t.Helper()
+
+	for _, rule := range jsonReplaceRules {
+		jsonInput = replaceJSONInput(t, jsonInput, rule.Path, rule.ReplaceFunc)
+	}
+
+	jsonFormatted := bytes.Buffer{}
+	err := json.Indent(&jsonFormatted, []byte(jsonInput), "", "  ")
+
+	if err != nil {
+		t.Fatalf("Failed to marshal JSON: %s", err)
+	}
+
+	return jsonFormatted.String()
 }
